Stop using 'terminate' package to avoid GPL
Fixes #212

1. The licensing issues

The 'terminate' package is licensed under GPL-2.0:
https://github.com/dwyl/terminate/blob/main/LICENSE

However, this og-image package is licensed under MIT.

Because 'terminate' was being used in the og-image *runtime* rather than as a dev
tool as the author of 'terminate' intended...:
  https://github.com/dwyl/terminate/issues/35#issuecomment-441572754
... this caused og-image to be a work derived from GPL-licensed code.

As per GPL-2.0 section 2 point B:

> https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html#section2
> You must cause any work that you distribute or publish, that in
whole or in part contains or is derived from the Program or any part
thereof, to be licensed as a whole at no charge to all third parties
under the terms of this License.

This essentially meant that if you use GPL code (here, it's
'terminate') as a part of your code (here, 'og-image'), your work MUST
be licensed under GPL as well.

However the intention of og-image is to still be licensed under MIT.

---

2. Replacing 'terminate'

Now, 'terminate' was first introduced in 'og-image' on September 6,
2023:
https://github.com/nuxt-modules/og-image/commit/e0cdde74906c5862757c070b379b393e2a9e8bd7#diff-030fc083b2cbf5cf008cfc0c49bb4f1b8d97ac07f93a291d068d81b4d1416f70R687

As Harlan said (https://github.com/nuxt-modules/og-image/issues/212#issuecomment-2149392241), 'terminate' was added to circumvent an issue where the
`npx playwright install chromium` process would hang and never exit.

Do note that at the time, the version of 'playwright' in the repo was
1.37.1:
https://github.com/nuxt-modules/og-image/blob/e0cdde74906c5862757c070b379b393e2a9e8bd7/package.json#L63

The issue where playwright doesn't exit after installing browsers
sounds like this issue:
https://github.com/microsoft/playwright/issues/28189

The issue was fixed on November 22, 2023 in these two commits:

- https://github.com/microsoft/playwright/commit/2f1b0d6ff751dec78e6f94175457422200cafcf1
- https://github.com/microsoft/playwright/commit/958e45316e0df430967d8fbf02676b3c4241a54a

We can see at the top of the UI that this commit is tagged 1.41.0.

At the time Harlan occured came across this issue the version was
again, 1.37.1.

So my guess it's that I could just revert to the previous way of
killing the process (from before September 6, 2023) and installation
is not going to freeze anymore, as the underlying Playwright issue is fixed.